EmployeeManagement/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
numpy==1.19.4
pytz==2020.4
six==1.15.0
sqlparse==0.4.1


